feat(logging): align addLogging for all applications
