+++ docker/Dockerfile-maintenance-service
@@ -28,6 +28,7 @@ COPY src/framework/Framework.BaseDependencies/ src/framework/Framework.BaseDepen
 COPY src/framework/Framework.DBAccess/ src/framework/Framework.DBAccess/
 COPY src/framework/Framework.Linq/ src/framework/Framework.Linq/
 COPY src/framework/Framework.Models/ src/framework/Framework.Models/
+COPY src/framework/Framework.Logging/ src/framework/Framework.Logging/
 COPY src/framework/Framework.ErrorHandling.Library/ src/framework/Framework.ErrorHandling.Library/
 RUN dotnet restore "src/maintenance/Maintenance.App/Maintenance.App.csproj"
 WORKDIR /src/maintenance/Maintenance.App
