+++ src/provisioning/Provisioning.Migrations/appsettings.json
@@ -5,6 +5,22 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
+  "Serilog": {
+    "Using":  [ "Serilog.Sinks.Console" ],
+    "MinimumLevel": "Information",
+    "WriteTo": [
+      { "Name": "Console" }
+    ],
+    "Enrich": [
+      "FromLogContext",
+      "WithMachineName",
+      "WithProcessId",
+      "WithThreadId"
+    ],
+    "Properties": {
+      "Application": "Org.Eclipse.TractusX.Portal.Backend.Provisioning.Migrations"
+    }
+  },
   "ConnectionStrings": {
     "ProvisioningDb": "Server=placeholder;Database=placeholder;Port=5432;User Id=placeholder;Password=placeholder;Ssl Mode=Disable;"
   },
