+++ tests/administration/Administration.Service.Tests/appsettings.IntegrationTests.json
@@ -1,4 +1,20 @@
 {
+  "Serilog": {
+    "Using":  [ "Serilog.Sinks.Console" ],
+    "MinimumLevel": "Warning",
+    "WriteTo": [
+      { "Name": "Console" }
+    ],
+    "Enrich": [
+      "FromLogContext",
+      "WithMachineName",
+      "WithProcessId",
+      "WithThreadId"
+    ],
+    "Properties": {
+      "Application": "Org.Eclipse.TractusX.Portal.Backend.Administration.Service.Tests"
+    }
+  },
   "Keycloak": {
     "central": {
       "ConnectionString": "",
