+++ tests/notifications/Notifications.Service.Tests/appsettings.IntegrationTests.json
@@ -6,6 +6,22 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
+  "Serilog": {
+    "Using":  [ "Serilog.Sinks.Console" ],
+    "MinimumLevel": "Warning",
+    "WriteTo": [
+      { "Name": "Console" }
+    ],
+    "Enrich": [
+      "FromLogContext",
+      "WithMachineName",
+      "WithProcessId",
+      "WithThreadId"
+    ],
+    "Properties": {
+      "Application": "Org.Eclipse.TractusX.Portal.Backend.Notifications.Service.Tests"
+    }
+  },
   "SwaggerEnabled": "",
   "HealthChecks": [],
   "Cors": {
